# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreatePersonRequest:
    properties:
      birth_date:
        type: string
      death_date:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        enum:
        - male
        - female
        type: string
      last_name:
        type: string
      notes:
        type: string
      phone:
        type: string
      photo:
        type: string
    required:
    - first_name
    - gender
    - last_name
    type: object
  handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: A family tree management API with authentication and role-based access.
  title: Family Tree API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses: {}
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses: {}
      summary: Login
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses: {}
      summary: Register a new user
      tags:
      - auth
  /persons:
    get:
      produces:
      - application/json
      responses: {}
      summary: List all persons
      tags:
      - persons
    post:
      consumes:
      - application/json
      parameters:
      - description: Person data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePersonRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new person
      tags:
      - persons
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"